requirements.txt
chainlit==1.3.2
langchain==1.1.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-groq==1.0.1
langchain-chroma==1.0.0
langchain-huggingface==1.0.1
langchain-text-splitters==1.0.0
langchain-classic==1.0.0 
python-dotenv==1.2.1
chromadb==1.3.5
FlashRank==0.2.10
rank-bm25==0.2.2
docx2txt==0.9
pydantic==2.10.1
sentence-transformers==5.1.2
torch==2.9.1